chore: rename compiler/proto to starproto
