+++ compiler/starproto/message_type.go
@@ -1,4 +1,4 @@
-package proto
+package starproto
 
 import (
 	"fmt"
